+++ frontend/packages/core/src/Surface/stories/card.stories.tsx
@@ -2,7 +2,7 @@ import * as React from "react";
 import type { Meta } from "@storybook/react";
 
 import type { CardProps } from "../card";
-import { Card, CardContent } from "../card";
+import { Card, CardContent } from "..";
 
 export default {
   title: "Core/Card/Basic",
