+++ frontend/packages/core/src/Surface/stories/paper.stories.tsx
@@ -1,7 +1,7 @@
 import * as React from "react";
 import type { Meta } from "@storybook/react";
 
-import Paper from "../paper";
+import { Paper } from "..";
 
 export default {
   title: "Core/Paper",
